Adding additional subtidal models from box
I believe I started working on integrating the subtidal model into the github.io website but it looks like it isn't all there. I'm uploading 2 files for this that I found on box. Again, I'm not certain that these are the most updated code, but it looks like they are after cross referencing what i was working on in creating the .qmd file for the website.